fix bugs of tape_t and relayout_view.
