+++ test/tensor.cpp
@@ -47,20 +47,22 @@ int main()
     auto vec = rzx::array{ 1.0, 2.0 };
     using vec_t = decltype(vec);    
 
-    //auto exp = rzx::mat_mul(mat, mat);
+    auto exp = rzx::mat_mul(mat, mat);
     //auto exp0 = rzx::mat_mul(mat0, mat0);
     //auto exp = rzx::dot(vec, vec);
     //std::cout << exp << "  gagaga\n";
     //auto exp = rzx::mat_mul_vec(mat, vec);
     //auto exp = mat | rzx::transpose<>;
-    auto exp = rzx::vec_mul_mat(vec, mat);
+    //auto exp = rzx::vec_mul_mat(vec, mat);
     //auto r = exp | rzx::make_tree<vec_t>;
-    //auto r = exp | rzx::make_tree<vec_t>;
+    auto r = exp | rzx::make_tree<mat_t>;
     //MAGIC_CHECK((mat | rzx::child<0, 0>), 1.0);
     //MAGIC_SHOW_TYPE(exp0);
     MAGIC_SHOW_TYPE(exp);
-    //MAGIC_SHOW_TYPE(r);
-    //std::cout << r[0] << ", " << r[1];
+    MAGIC_SHOW_TYPE(r);
+    
+    std::cout << r[0][0] << ", " << r[0][1] << '\n';
+    std::cout << r[1][0] << ", " << r[1][1] << '\n';
     //std::array<std::array<double, 2>, 2> e{};
     //rzx::mat_mul(r, mat) >> e;
